Refactor alocation context imports
