+++ src/state/index.js
@@ -1,4 +1,6 @@
 import { Context as AuthContext } from './AuthContext.js';
 import { Provider as AuthProvider } from './AuthContext';
+import { Provider as LocationProvider } from './LocationContext';
+import { Context as LocationContext } from './LocationContext';
 
-export { AuthContext, AuthProvider };
+export { AuthContext, AuthProvider, LocationProvider, LocationContext };
